requirements.txt
flask==1.1.1
flask-sqlalchemy==2.4.1
sqlalchemy==1.3.12
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0
pymysql==0.9.3
flask-injector==0.12.1
injector==0.18.2
flask-jwt-extended==3.24.1
flask-migrate==2.5.2
python-dotenv==0.10.3
bcrypt==3.1.7
pandas==0.25.3
xlsxwriter==1.2.7
redis==3.3.11
celery==4.4.0
flask-celery==2.4.3
kombu==4.6.7
CeleryFlask>= 0.0.1
